requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.2.7
django-allauth==0.57.0
django-cors-headers==4.6.0
django-email-verification==0.3.3
django-jazzmin==3.0.1
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
elevenlabs==1.57.0
environ==1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
google-auth==2.40.3
google-genai==1.26.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
oauthlib==3.3.1
openai==1.97.0
openpyxl==3.1.2
packaging==25.0
payme-pkg==3.0.24
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-environ==0.4.54
python3-openid==3.2.0
qrcode==7.4.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
validators==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.11.0
